Add bold italic support, major refactor
- *text* is now parsed as italic text, while **text** is bold
- The major refactor includes more documentation and testing,
and re-doing the text tokens. Text now is blob of data that isn't
directly translated to HTML. It is contained in a parent structure
paragraph, which is essentially just a single markdown line. It's just
called paragraph since it matches the HTML output.
A paragraph (line) is a vector of texts, and is used in the heading as
well. This allows a line to have styled portions (bold/italic)